requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
common==0.1.2
Django==3.1.1
djangorestframework==3.11.1
googletrans==3.0.0
gTTS==2.1.1
gTTS-token==1.1.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.9.9
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
langdetect==1.0.8
lxml==4.5.2
psycopg2==2.8.5
python-vlc==3.0.11115
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rfc3986==1.4.0
selenium==3.141.0
six==1.15.0
sniffio==1.1.0
soupsieve==2.0.1
sqlparse==0.3.1
translation==1.0.5
urllib3==1.25.10
